refactor(settings): move destructive actions to Account, drop Connections
- Extract Log Out / Clear App Data + the confirmation modal out of
  Settings home into a new LogoutAndClearActions component and render
  it as the footer of the Settings → Account page. SettingsSectionPage
  grows an optional `footer` slot. The home menu is now purely
  navigational; destructive actions live with the rest of the
  account-scoped controls.
- Delete the Connections panel entirely (MetaMask / Binance / Notion /
  Google placeholders that are no longer part of the product), along
  with its test and account-section entry. The deep route
  `/settings/connections` is gone; references in the navigation hook,
  Developer Options, and the post-onboarding deep link are removed or
  repointed to `/settings/composio-routing`. `walletApi` stays — it's
  still used by the recovery-phrase flow.
- Move the Clear-App-Data flow tests to the new component, and add a
  guard on SettingsHome that the destructive entries no longer appear
  on the home screen.
